pyproject.toml
[project]
name = "playwright-api-web-mobile"
version = "0.1.0"
description = "Production-grade QA automation showcase using Practice Software Testing demo app"
authors = [
    {name = "Kenny Ramadhan", email = "kennyrmdhn@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["qa", "automation", "playwright", "pytest", "allure", "portfolio"]
dependencies = [
    "playwright>=1.49.0",
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-rerunfailures>=14.0",
    "pytest-xdist>=3.6.0",
    "allure-pytest>=2.13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",
    "faker>=30.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--alluredir=reports/allure-results",
    "--clean-alluredir",
    "--tb=short",
]
markers = [
    "smoke: smoke tests (fast, run on every PR)",
    "regression: full regression suite",
    "slow: known slow tests",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "RET", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_decorators = false
